keerthi_kumar_assignment/2nd question/source_code/orm_code/emp_mgr/requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
